refactor: move slipday related enrollment methods to separate file
